INTERLOK-3294 Upgrade cassandra to 4.17.0 and upgrade the code to use the new API
The Cassandra has now the possibility to set the port via the connectUrl: hostname:port. If the port is not provided the default port '9042' will be used.
It also has a new option to set the datacenter name which is now required by the new API. If left blank the considered default 'datacenter1' will be used.